move validators.py from users to core
